requirements.txt
torch>=1.10.0
torchvision>=0.11.0
numpy
pandas
scikit-learn
matplotlib
seaborn
opencv-python
opencv-contrib-python
pillow
grad-cam
tensorboard
tqdm
ultralytics>=8.0.0
scipy
scikit-image
